+++ R/dependencies.R
@@ -12,17 +12,17 @@
 #'   be character of the form `openpharma/stageddeps.food@https://github.com`
 #'   If host is not included in the string then the default `https://github.com`
 #'   is assumed.
-#' @param project_type (`character`) See `project` argument
+#' @param project_type (`character`) See `project` argument.
 #' @param ref (`character`) git branch (or tag) inferred from the
 #'   branch of the project if not provided; warning if not consistent with
 #'   current branch of project. If `project_type` is not `local` then this argument
-#'   must be provided
+#'   must be provided.
 #' @param local_repos (`data.frame`) repositories that should be taken from
-#'   local file system rather than cloned; columns are `repo, host, directory`
+#'   local file system rather than cloned; columns are `repo, host, directory`.
 #' @param direction (`character`) direction in which to discover packages
 #'   either "upstream","downstream" or "all".
 #' @param fallback_branch (`character`) the default branch to try to use if
-#'   no other matches found
+#'   no other matches found. It defaults to `"main"`.
 #' @param renv_profile (`character`) the name of the renv profile of the `renv.lock` files
 #'   to be included from the repos. The standard `renv.lock` file uses the default `NULL` argument here.
 #' @inheritParams argument_convention
@@ -47,8 +47,6 @@
 #'   \item{renv_files}{`named list` containing the json of the renv.lock files for the chosen profile for
 #'                     each repo. An entry to the list is `NULL` if a repos does not have the required lock file}
 #' }
-#' @md
-#' @export
 #' @examples
 #' \dontrun{
 #' dependency_table(verbose = 1)
@@ -64,6 +62,7 @@
 #' print(x)
 #' plot(x)
 #' }
+#' @export
 dependency_table <- function(project = ".",
                              project_type = c("local", "repo@host")[1],
                              ref = NULL,
@@ -75,25 +74,27 @@ dependency_table <- function(project = ".",
 
   # validate arguments
   stopifnot(is.data.frame(local_repos) || is.null(local_repos))
-  check_verbose_arg(verbose)
   direction <- check_direction_arg_deprecated(direction)
   check_direction_arg(direction)
   stopifnot(project_type %in% c("local", "repo@host"))
   stopifnot(rlang::is_scalar_character(fallback_branch))
 
+  if (verbose != 1) {
+    checkmate::assert_int(verbose, lower = 0, upper = 2)
+    verbose_sd_set(verbose)
+  }
+
   if (project_type == "repo@host" && (is.null(ref) || nchar(ref) == 0)) {
     stop("For non-local projects the (branch/tag) must be specified")
   }
   if (project_type == "local") {
-    check_dir_exists(project)
+    checkmate::assert_directory_exists(project)
     error_if_stageddeps_inexistent(project)
     # infer ref if not given
     if (is.null(ref) || nchar(ref) == 0) {
       ref <- infer_ref_from_branch(project)
     }
-  }
 
-  if (project_type == "local") {
     # take local version of project (rather than remote)
     local_repos <- add_project_to_local_repos(project, local_repos)
     repo_deps_info <- get_yaml_deps_info(project)
@@ -105,7 +106,7 @@ dependency_table <- function(project = ".",
       ),
       fallback_branch
     )
-    repo_to_process <- list(repo_deps_info$current_repo)
+    repo_to_process <- list(repo_deps_info$current_repo) # This list will be updated with the deps
   } else {
     repo_to_process <- list(parse_remote_project(project))
   }
@@ -115,8 +116,7 @@ dependency_table <- function(project = ".",
   internal_deps <- rec_checkout_internal_deps(
     repo_to_process, ref,
     direction = direction,
-    local_repos = local_repos, fallback_branch = fallback_branch,
-    verbose = verbose
+    local_repos = local_repos, fallback_branch = fallback_branch
   )
 
   internal_deps$package_name[internal_deps$accessible] <-
@@ -216,7 +216,7 @@ print.dependency_structure <- function(x, ...) {
 
   if (!all(table$installable)) {
     table$package_name <- paste0(table$package_name, ifelse(table$installable, "", "*"))
-    cat(
+    warning(
       "packages denoted with '*' cannot be installed as either they or one of their internal",
       "dependencies is not accessible\n"
     )
@@ -353,7 +353,6 @@ plot.dependency_structure <- function(x, y, ...) {
 #'
 #' @return `data.frame` of performed actions
 #'
-#' @export
 #' @seealso determine_branch
 #'
 #' @examples
@@ -365,6 +364,7 @@ plot.dependency_structure <- function(x, y, ...) {
 #' # install all dependencies
 #' install_deps(x, install_direction = "all")
 #' }
+#' @export
 install_deps <- function(dep_structure,
                          install_project = TRUE,
                          install_direction = "upstream",
@@ -380,6 +380,11 @@ install_deps <- function(dep_structure,
   install_direction <- check_direction_arg_deprecated(install_direction)
   if (dep_structure$direction != "all" && dep_structure$direction != install_direction) {
     stop("Invalid install_direction argument for this dependency object")
+  }
+
+  if (verbose != 1) {
+    checkmate::assert_int(verbose, lower = 0, upper = 2)
+    verbose_sd_set(verbose)
   }
 
   # get the packages to install
@@ -402,8 +407,9 @@ install_deps <- function(dep_structure,
     install_external_deps = install_external_deps,
     upgrade = upgrade,
     internal_pkg_deps = dep_structure$table$package_name,
-    verbose = verbose, ...
+    ...
   )
+
   pkg_actions
 }
 
@@ -447,6 +453,11 @@ check_downstream <- function(dep_structure,
     stop("Invalid dependency table - downstream dependencies must be have been calculated")
   }
 
+  if (verbose != 1) {
+    checkmate::assert_int(verbose, lower = 0, upper = 2)
+    verbose_sd_set(verbose)
+  }
+
   # get the packages to install
   pkg_df <- dep_structure$table
 
@@ -471,7 +482,7 @@ check_downstream <- function(dep_structure,
     upgrade = upgrade,
     internal_pkg_deps = dep_structure$table$package_name,
     rcmd_args = list(check = check_args),
-    verbose = verbose, ...
+    ...
   )
 
   pkg_actions
@@ -507,9 +518,7 @@ update_with_direct_deps <- function(dep_structure) {
 
 #' Build, check and install internal dependencies
 #'
-
 #'
-#' @md
 #' @inheritParams install_deps
 #' @inheritDotParams install_deps
 #' @param steps (`character` vector) subset of "build", "check", "install";
@@ -522,13 +531,13 @@ update_with_direct_deps <- function(dep_structure) {
 #' @return list with entries
 #'  - artifact_dir: `artifact_dir` directory with log files
 #'  - pkg_actions: `data.frame` of performed actions
-#' @export
 #' @examples
 #' \dontrun{
 #' x <- dependency_table(project = ".", verbose = 1)
 #' build_check_install(x, steps = c("build", "check"), verbose = 1)
 #' build_check_install(x, artifact_dir = "../output")
 #' }
+#' @export
 build_check_install <- function(dep_structure,
                                 install_direction = "all",
                                 steps = c("build", "check", "install"),
@@ -544,6 +553,10 @@ build_check_install <- function(dep_structure,
     methods::is(dep_structure, "dependency_structure"),
     is.logical(dry)
   )
+  if (verbose != 1) {
+    checkmate::assert_int(verbose, lower = 0, upper = 2)
+    verbose_sd_set(verbose)
+  }
 
   install_direction <- check_direction_arg_deprecated(install_direction)
   if (dep_structure$direction != "all" && dep_structure$direction != install_direction) {
@@ -577,7 +590,7 @@ build_check_install <- function(dep_structure,
     internal_pkg_deps = dep_structure$table$package_name,
     rcmd_args = rcmd_args,
     artifact_dir = artifact_dir,
-    verbose = verbose, ...
+    ...
   )
 
   return(list(artifact_dir = artifact_dir, pkg_actions = pkg_actions))
@@ -674,7 +687,6 @@ check_yamls_consistent <- function(dep_structure, skip_if_missing_yaml = FALSE) 
 #'   which fields of the DESCRIPTION file of the internal packages should be included. Default:
 #'   `c("Depends", "Imports", "LinkingTo")`
 #' @inheritParams build_check_install
-#' @md
 #' @return A vector of 'external' R packages required to install
 #'   the selected 'internal' packages, ordered by install order (unless `from_external_dependencies`
 #'   does not include `"Depends"`, `"Imports"` and `"LinkingTo"`). The core R packages