+++ R/dependencies_helper.R
@@ -24,7 +24,6 @@
 #'   `build`; action `test` only outputs to the console
 #' @param ... Additional args passed to [remotes::install_deps()] Ignored
 #'   if inside an `renv` environment.
-#' @inheritParams argument_convention
 #'
 #' @keywords internal
 run_package_actions <- function(pkg_actions, internal_pkg_deps,
@@ -33,15 +32,14 @@ run_package_actions <- function(pkg_actions, internal_pkg_deps,
                                 upgrade = "never",
                                 rcmd_args = NULL,
                                 artifact_dir = NULL,
-                                verbose = 0, ...) {
+                                ...) {
   all_actions <- unique(unlist(pkg_actions$actions))
   stopifnot(all(all_actions %in% c("test", "build", "check", "install")))
-  check_verbose_arg(verbose)
 
   rcmdcheck_outputs <- list()
 
   if (nrow(pkg_actions) == 0) {
-    message_if_verbose("No packages to process!", verbose = verbose)
+    message_if_verbose("No packages to process!")
     return(pkg_actions)
   }
 
@@ -58,8 +56,7 @@ run_package_actions <- function(pkg_actions, internal_pkg_deps,
   }
 
   message_if_verbose("Processing packages in order: ",
-    toString(pkg_actions$package_name),
-    verbose = verbose
+    toString(pkg_actions$package_name)
   )
 
   for (idx in seq_along(pkg_actions$cache_dir)) {
@@ -68,8 +65,7 @@ run_package_actions <- function(pkg_actions, internal_pkg_deps,
 
     if (!pkg_actions$installable[idx]) {
       message_if_verbose("skipping package ", pkg_actions$package_name[idx],
-        " as it (or one of its upstream dependencies) is not accessible",
-        verbose = verbose
+        " as it (or one of its upstream dependencies) is not accessible"
       )
       next
     }
@@ -106,7 +102,7 @@ run_package_actions <- function(pkg_actions, internal_pkg_deps,
             stop("Tests for package in directory ", cache_dir, " failed")
           }
         } else {
-          message_if_verbose("No tests found for package in directory ", cache_dir, verbose = verbose)
+          message_if_verbose("No tests found for package in directory ", cache_dir)
         }
       }
 
@@ -160,11 +156,11 @@ run_package_actions <- function(pkg_actions, internal_pkg_deps,
         }
       }
     } else { # dry run
-      message_if_verbose(cat_nl("(Dry run) Skipping", toString(actions), "of", cache_dir), verbose = verbose)
+      message_if_verbose(cat_nl("(Dry run) Skipping", toString(actions), "of", cache_dir))
     }
   }
 
-  message_if_verbose("Processed packages in order: ", toString(pkg_actions$package_name), verbose = verbose)
+  message_if_verbose("Processed packages in order: ", toString(pkg_actions$package_name))
 
   # output rcmdcheck outputs
   if (!rlang::is_empty(rcmdcheck_outputs)) {
@@ -187,7 +183,7 @@ add_project_to_local_repos <- function(project, local_repos) {
   stopifnot(
     is.data.frame(local_repos) || is.null(local_repos)
   )
-  check_dir_exists(project)
+  checkmate::assert_directory_exists(project)
 
   repo_deps_info <- get_yaml_deps_info(project)
   rbind(